requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
pgvector==0.2.5
sentence-transformers==3.0.1
torch>=2.2,<2.9
transformers>=4.41.0
huggingface-hub>=0.23.4
spacy==3.7.4
pymupdf==1.24.5
pydantic==2.7.3
python-multipart==0.0.9
numpy==1.24.4
